Added assigned buildings raport
