+++ projekt/Budowa/src/main/java/org/budowa/flow/shared/DashboardBaseController.java
@@ -9,6 +9,7 @@ import org.budowa.entities.BuildingStatus;
 import org.budowa.flow.kanban.KanbanController;
 import org.budowa.router.Router;
 import org.budowa.services.DialogService;
+import org.budowa.services.PdfBuilder;
 import org.budowa.texts.Translations;
 
 import java.io.IOException;
@@ -20,6 +21,7 @@ public abstract class DashboardBaseController implements Initializable {
 
     private final Router router = Router.inject();
     private final DialogService dialogService = DialogService.inject();
+    private final PdfBuilder pdfBuilder = PdfBuilder.inject();
 
     @FXML
     public KanbanController kanbanController;
@@ -63,5 +65,22 @@ public abstract class DashboardBaseController implements Initializable {
             }
         });
         return label;
+    }
+
+    protected void printRaport(String title) {
+        try {
+            var buildings = this.loadBuildings();
+
+            var builder = this.pdfBuilder.create(title).addText(title).addEmptyLine();
+            for (var building : buildings) {
+                builder = builder.addDataBlock("Nazwa:", building.getName())
+                        .addDataBlock("Status:", building.getStatus().toString())
+                        .addEmptyLine();
+            }
+            builder.save();
+            this.dialogService.showInfoDialog(Translations.SUCCESSFULLY_SAVED_PDF);
+        } catch (Exception ex) {
+            this.dialogService.showErrorDialog(Translations.SOMETHING_WENT_WRONG);
+        }
     }
 }
