+++ projekt/Budowa/src/main/java/org/budowa/texts/Translations.java
@@ -3,4 +3,5 @@ package org.budowa.texts;
 public class Translations {
     public static String SOMETHING_WENT_WRONG = "Coś poszło nie tak";
     public static String SUCCESSFULLY_SAVED_PDF = "Pomyślnie zapisano PDF.";
+    public static String ASSIGNED_BUILDINGS = "Przypisane budowy";
 }
